IngestionServices/requirements.txt
google-cloud-storage
pika
pandas